# PortfolioApp/PortfolioApp.Application/DTOs/Dtos.cs
using PortfolioApp.Application.Strategies;
using PortfolioApp.Domain;

namespace PortfolioApp.Application.DTOs
{
    public record BuyRequestDto(Guid? AssetId, string? Ticker, int Qty, decimal Price, DateTime Date);

    public record SellRequestDto(
        Guid? AssetId,
        string? Ticker,
        int Qty,
        decimal Price,
        CostBasisMethod Method,
        DateTime Date);

    /// <param name="LastUnitCost">Unit cost of the most recently bought lot.</param>
    public record AssetDto(string Ticker, int QtyRemaining, decimal LastUnitCost);
    public record LotDto(
        Guid Id,
        string Ticker,
        DateTime PurchaseDate,
        int QtyInitial,
        int QtyRemain,
        decimal RawUnitCost,
        decimal UnitCost);

    public record TradeDto(
        Guid Id,
        string Ticker,
        DateTime Date,
        int Quantity,
        decimal Price,
        decimal ProfitGross,
        decimal ProfitNet);

    public record FeeRuleDto(FeeType Type, decimal Amount, FeeDirection Direction);

    public static class FeeRuleMapping
    {
        public static FeeRuleDto ToDto(this FeeRegistration r) => new(r.Type, r.Amount, r.Direction);

        public static FeeRegistration ToDomain(this FeeRuleDto d) => new(d.Type, d.Amount, d.Direction);
    }
}