# PortfolioApp/portfolio-ui/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Portfolio API",
    "version": "v1"
  },
  "paths": {
    "/api/fees/get": {
      "get": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeeRuleDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/fees/save": {
      "post": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FeeRuleDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/fees/delete": {
      "delete": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/assets/id/{ticker}": {
      "get": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets": {
      "get": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trades/buy": {
      "post": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/trades/sell": {
      "post": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/lots": {
      "get": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LotDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trades/all": {
      "get": {
        "tags": [
          "PortfolioApp.Web"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradeDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssetDto": {
        "required": [
          "lastUnitCost",
          "qtyRemaining",
          "ticker"
        ],
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "qtyRemaining": {
            "type": "integer",
            "format": "int32"
          },
          "lastUnitCost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "BuyRequestDto": {
        "required": [
          "date",
          "price",
          "qty"
        ],
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ticker": {
            "type": "string",
            "nullable": true
          },
          "qty": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CostBasisMethod": {
        "enum": [
          "FIFO",
          "LIFO",
          "Average"
        ],
        "type": "string"
      },
      "FeeDirection": {
        "enum": [
          "None",
          "Buy",
          "Sell",
          "Both"
        ],
        "type": "string"
      },
      "FeeRuleDto": {
        "required": [
          "amount",
          "direction",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/FeeType"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "direction": {
            "$ref": "#/components/schemas/FeeDirection"
          }
        },
        "additionalProperties": false
      },
      "FeeType": {
        "enum": [
          "FixedPerTrade",
          "FixedPerShare",
          "Percent"
        ],
        "type": "string"
      },
      "LotDto": {
        "required": [
          "id",
          "purchaseDate",
          "qtyInitial",
          "qtyRemain",
          "rawUnitCost",
          "ticker",
          "unitCost"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ticker": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "qtyInitial": {
            "type": "integer",
            "format": "int32"
          },
          "qtyRemain": {
            "type": "integer",
            "format": "int32"
          },
          "rawUnitCost": {
            "type": "number",
            "format": "double"
          },
          "unitCost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "SaleResult": {
        "required": [
          "grossProfit",
          "netProfit",
          "remainingCostPerShare",
          "remainingShares",
          "soldCostPerShare"
        ],
        "type": "object",
        "properties": {
          "remainingShares": {
            "type": "integer",
            "format": "int32"
          },
          "soldCostPerShare": {
            "type": "number",
            "format": "double"
          },
          "remainingCostPerShare": {
            "type": "number",
            "format": "double"
          },
          "grossProfit": {
            "type": "number",
            "format": "double"
          },
          "netProfit": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "SellRequestDto": {
        "required": [
          "date",
          "method",
          "price",
          "qty"
        ],
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ticker": {
            "type": "string",
            "nullable": true
          },
          "qty": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "method": {
            "$ref": "#/components/schemas/CostBasisMethod"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "TradeDto": {
        "required": [
          "date",
          "id",
          "price",
          "profitGross",
          "profitNet",
          "quantity",
          "ticker"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ticker": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "profitGross": {
            "type": "number",
            "format": "double"
          },
          "profitNet": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}